requirements.txt
# Core ML Libraries
torch>=1.13.0
transformers>=4.30.2
datasets>=2.13.2
accelerate>=0.20.3
peft>=0.3.0

# API Framework
fastapi>=0.102.0
pydantic>=2.4.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data Processing
pandas>=1.3.5
numpy>=1.21.6
scikit-learn>=1.0.2

# Monitoring & Logging
wandb>=0.16.0
tensorboard>=2.11.2

# Utilities
python-dotenv>=0.21.1
requests>=2.31.0
beautifulsoup4>=4.12.0
tqdm>=4.66.0

# Development
pytest>=7.4.0
black>=23.3.0
flake8>=5.0.4
pre-commit>=2.21.0

# Cloud & Deployment
google-cloud-storage>=2.10.0
docker>=6.1.0

